tools/web_search_tool.py
from langchain.tools import tool
# import requests
# from bs4 import BeautifulSoup
from tavily import TavilyClient
import os
from dotenv import load_dotenv
from rich import print

load_dotenv()

tavily=TavilyClient(api_key=os.getenv("TAVILY_API_KEY"))

@tool
def web_search(query : str) -> str:
    """
    Searches the web for recent and reliable information on a topic. Returns titles, urls and snippets.
    """

    results=tavily.search(query=query, max_results=5)

    out=[]

    for r in results['results']:
        out.append(
            f"Title: {r['title']}\nURL: {r['url']}\nSnippet: {r['content'][:300]}\n"
        )
    

    # print("\n------\n".join(out))

    return "\n------\n".join(out)

