requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
duckduckgo_search==8.1.1
faiss-cpu==1.13.2
ffmpeg-python==0.2.0
filelock==3.25.2
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==25.12.19
fonttools==4.62.1
fpdf2==2.8.7
frozenlist==1.8.0
fsspec==2026.2.0
future==1.0.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.47.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.3.2
groq==0.37.1
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.7.2
idna==3.11
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==2.12.0
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.28
langchain-groq==1.1.2
langchain-huggingface==1.2.1
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.7.30
libclang==18.1.1
lxml==6.0.3
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.15.0
networkx==3.6.1
numpy==2.4.3
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==24.2
pandas==2.3.3
pillow==12.1.0
plotly==6.6.0
primp==1.2.2
propcache==0.4.1
protobuf==4.25.8
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.9.2
pydantic-settings==2.13.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
reportlab==4.4.10
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.15.3
sentence-transformers==5.3.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
streamlit==1.52.2
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.11.0
tornado==6.5.4
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uv==0.11.7
uvicorn==0.44.0
watchdog==6.0.0
Werkzeug==3.1.4
wrapt==1.14.2
xxhash==3.6.0
yarl==1.23.0
yt-dlp==2026.3.17
zstandard==0.25.0
